requirements.txt
Django==4.1.2
djangorestframework==3.14.0
PyMySQL==1.0.2
pymongo==3.12.1
djongo==1.3.6
